v1.9 - 'fixed' math cmds and reworked images and thumbnails

+++ commands/divide.js
@@ -11,10 +11,10 @@ module.exports = {
 
         switch(primaryCommand.toLowerCase()) { // called when messages says 'pog divide'
             case 'divide' :
-                var quotient = parseInt(first) / parseInt(second) // divide the first number by the second number
+                var quotient = parseFloat(first) / parseFloat(second) // divide the first number by the second number
                 const embed = new MessageEmbed()
                 .setTitle(quotient)
-                .setDescription("<@!" + receivedMessage.author.id + ">" + `\n${parseInt(first)} ÷ ${parseInt(second)} =`)
+                .setDescription("<@!" + receivedMessage.author.id + ">" + `\n${parseFloat(first)} ÷ ${parseFloat(second)} =`)
                 .setColor('#00ADEF')
                 receivedMessage.channel.send(embed) // sends the quotient in the chat
         }
